4 heatmap provider

+++ island/src/main/java/ca/mcmaster/cas/se2aa4/a3/island/shape/CircleShape.java
@@ -1,18 +1,18 @@
 package ca.mcmaster.cas.se2aa4.a3.island.shape;
 
 public class CircleShape implements Shape {
-    private final int x;
-    private final int y;
-    private final int radius;
+    private final double x;
+    private final double y;
+    private final double radius;
 
-    public CircleShape(int x, int y, int radius) {
+    public CircleShape(double x, double y, double radius) {
         this.x = x;
         this.y = y;
         this.radius = radius;
     }
 
     @Override
-    public boolean contains(int x, int y) {
+    public boolean contains(double x, double y) {
         return Math.pow(x - this.x, 2) + Math.pow(y - this.y, 2) <= Math.pow(radius, 2);
     }
 }
